% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCell_functions.R
\name{dCell.expProcessing}
\alias{dCell.expProcessing}
\title{Pre-process RNASeq based expression data.
\code{dCell.expProcessing} Performes a TMM normalization, log2(counts+1) and scaling.}
\usage{
dCell.expProcessing(count.table, trim = TRUE)
}
\arguments{
\item{count.table}{A table of quantified gene expression, samples by columns and genes by row.}

\item{trim}{boolean, whather the expression data should be trimmed to include only genes present in DeconCell models.}
}
\value{
A normalized and scaled expression matrix.
}
\description{
Pre-process RNASeq based expression data.
\code{dCell.expProcessing} Performes a TMM normalization, log2(counts+1) and scaling.
}
\examples{
data(coun.table)
dCell.exp <- dCell.expProcessing(count.table)
dim(dCell.exp)
head(dCell.exp)
}
